refactor: eliminate duplication of Rho type definitions
