+++ casper/src/test/scala/coop/rchain/casper/helper/RhoLoggerContract.scala
@@ -2,7 +2,7 @@ package coop.rchain.casper.helper
 import cats.effect.Sync
 import coop.rchain.models.ListParWithRandomAndPhlos
 import coop.rchain.rholang.interpreter.Runtime.SystemProcess
-import coop.rchain.rholang.interpreter.{ContractCall, PrettyPrinter}
+import coop.rchain.rholang.interpreter.{ContractCall, PrettyPrinter, RhoType}
 import coop.rchain.shared.{Log, LogSource}
 
 object RhoLoggerContract {
@@ -14,7 +14,7 @@ object RhoLoggerContract {
     val isContractCall = new ContractCall(ctx.space, ctx.dispatcher)
 
     message match {
-      case isContractCall(_, Seq(IsString(logLevel), par)) =>
+      case isContractCall(_, Seq(RhoType.String(logLevel), par)) =>
         val msg         = prettyPrinter.buildString(par)
         implicit val ev = LogSource.matLogSource
