+++ casper/src/test/scala/coop/rchain/casper/helper/TestResultCollector.scala
@@ -3,43 +3,24 @@ import cats.effect.Sync
 import cats.effect.concurrent.Ref
 import cats.implicits._
 import coop.rchain.crypto.hash.Blake2b512Random
-import coop.rchain.models.Expr.ExprInstance.{ETupleBody, GBool, GInt, GString}
+import coop.rchain.models.Expr.ExprInstance.{ETupleBody, GBool}
 import coop.rchain.models.rholang.implicits._
-import coop.rchain.models.{ETuple, Expr, ListParWithRandomAndPhlos, Par, _}
-import coop.rchain.rholang.interpreter.ContractCall
+import coop.rchain.models.{ETuple, Expr, ListParWithRandomAndPhlos, Par}
 import coop.rchain.rholang.interpreter.Runtime.SystemProcess
-import coop.rchain.rholang.interpreter.accounting.Cost
-import coop.rchain.rholang.interpreter.errors.OutOfPhlogistonsError
-import coop.rchain.rholang.interpreter.storage.implicits.matchListPar
-import coop.rchain.rspace.util.unpackCont
-
-object IsString {
-  def unapply(p: Par): Option[String] =
-    p.singleExpr().collect {
-      case Expr(GString(bs)) => bs
-    }
-}
-
-object IsBoolean {
-  def unapply(p: Par): Option[Boolean] =
-    p.singleExpr().collect {
-      case Expr(GBool(b)) => b
-    }
-}
-
-object IsNumber {
-  def unapply(p: Par): Option[Long] =
-    p.singleExpr().collect {
-      case Expr(GInt(v)) => v
-    }
-}
+import coop.rchain.rholang.interpreter.{ContractCall, RhoType}
 
 object IsAssert {
   def unapply(
       p: Seq[Par]
   ): Option[(String, Long, Par, String, Par)] =
     p match {
-      case Seq(IsString(testName), IsNumber(attempt), assertion, IsString(clue), ackChannel) =>
+      case Seq(
+          RhoType.String(testName),
+          RhoType.Number(attempt),
+          assertion,
+          RhoType.String(clue),
+          ackChannel
+          ) =>
         Some((testName, attempt, assertion, clue, ackChannel))
       case _ => None
     }
@@ -50,14 +31,14 @@ object IsComparison {
       p: Par
   ): Option[(Par, String, Par)] =
     p.singleExpr().collect {
-      case Expr(ETupleBody(ETuple(List(expected, IsString(operator), actual), _, _))) =>
+      case Expr(ETupleBody(ETuple(List(expected, RhoType.String(operator), actual), _, _))) =>
         (expected, operator, actual)
     }
 }
 object IsSetFinished {
   def unapply(p: Seq[Par]): Option[Boolean] =
     p match {
-      case Seq(IsBoolean(hasFinished)) =>
+      case Seq(RhoType.Boolean(hasFinished)) =>
         Some(hasFinished)
       case _ => None
     }
@@ -117,21 +98,19 @@ class TestResultCollector[F[_]: Sync](result: Ref[F, TestResult]) {
     message match {
       case isContractCall(produce, IsAssert(testName, attempt, assertion, clue, ackChannel)) =>
         assertion match {
-          case IsComparison(expected, "==", actual) => {
+          case IsComparison(expected, "==", actual) =>
             val assertion = RhoAssertEquals(testName, expected, actual, clue)
             for {
               _ <- result.update(_.addAssertion(attempt, assertion))
               _ <- produce(Seq(Expr(GBool(assertion.isSuccess))), ackChannel)
             } yield ()
-          }
-          case IsComparison(unexpected, "!=", actual) => {
+          case IsComparison(unexpected, "!=", actual) =>
             val assertion = RhoAssertNotEquals(testName, unexpected, actual, clue)
             for {
               _ <- result.update(_.addAssertion(attempt, assertion))
               _ <- produce(Seq(Expr(GBool(assertion.isSuccess))), ackChannel)
             } yield ()
-          }
-          case IsBoolean(condition) =>
+          case RhoType.Boolean(condition) =>
             for {
               _ <- result.update(_.addAssertion(attempt, RhoAssertTrue(testName, condition, clue)))
               _ <- produce(Seq(Expr(GBool(condition))), ackChannel)
