+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/RhoType.scala
@@ -1,11 +1,12 @@
 package coop.rchain.rholang.interpreter
 import com.google.protobuf.ByteString
-import coop.rchain.models.Expr.ExprInstance.{GBool, GByteArray, GString}
+import coop.rchain.models.Expr.ExprInstance.{GBool, GByteArray, GInt, GString}
 import coop.rchain.models.{Expr, Par}
 
 object RhoType {
+  import coop.rchain.models.rholang.implicits._
+
   object ByteArray {
-    import coop.rchain.models.rholang.implicits._
     def unapply(p: Par): Option[Array[Byte]] =
       p.singleExpr().collect {
         case Expr(GByteArray(bs)) => bs.toByteArray
@@ -24,7 +25,20 @@ object RhoType {
     def apply(s: String): Par = GString(s)
   }
 
-  object Bool {
+  object Boolean {
     def apply(b: Boolean) = Expr(GBool(b))
+
+    def unapply(p: Par): Option[Boolean] =
+      p.singleExpr().collect {
+        case Expr(GBool(b)) => b
+      }
   }
+
+  object Number {
+    def unapply(p: Par): Option[Long] =
+      p.singleExpr().collect {
+        case Expr(GInt(v)) => v
+      }
+  }
+
 }
